+++ types/resources/OmnichannelSubscription.d.ts
@@ -9,7 +9,8 @@ declare module 'chargebee' {
     source: 'apple_app_store';
     customer_id?: string;
     created_at: number;
-    omnichannel_subscription_items: OmnichannelSubscription.OmnichannelSubscriptionItem[];
+    resource_version?: number;
+    omnichannel_subscription_items: OmnichannelSubscriptionItem[];
   }
 
   export namespace OmnichannelSubscription {
@@ -47,25 +48,13 @@ declare module 'chargebee' {
       next_offset?: string;
     }
 
-    export interface OmnichannelSubscriptionItem {
-      id: string;
-      id_at_source: string;
-      status: 'active' | 'expired' | 'cancelled';
-      current_term_start?: number;
-      current_term_end?: number;
-      expired_at?: number;
-      expiration_reason?: 'billing_error' | 'product_not_available' | 'other';
-      cancelled_at?: number;
-      cancellation_reason?:
-        | 'customer_cancelled'
-        | 'customer_did_not_consent_to_price_increase';
-    }
     // REQUEST PARAMS
     //---------------
 
     export interface ListInputParam {
       limit?: number;
       offset?: string;
+      customer_id?: filter.String;
     }
     export interface OmnichannelTransactionsForOmnichannelSubscriptionInputParam {
       limit?: number;